completion of part 2 and partial of part 3
I used different fonctions in part 2 to better understand our dataset and what we want to look at more in details.
For part three it is a deeper analysis of each variable present in  our dataset and looking at their significance. As well as their relationships within the sample set.