corrections in part 3 numerical analysis
I replaced my describe fonction with the individual fonctions so that i can see my results more cleanly and grouped together. I have not yet figured out how to or if i can use a single line of code insated of having to repeat over and over. But what i have done still gives us the correct results we are looking for.
Also found the number of unique categories for my categorical columns .which complete the manipulations that needed to be done with them.