changes to my conditional
i added the value of the skew that was calculated with the item term. However ive been trying to include the name of the variable that the conditional and the value of skew applies to so that we can understand the results more clearly. im having difficulty figuring out how to achieve this.