found how to print the variable name
when i did research and played around with the data i found that the most effective way to print both the name and value of the skew is via a dictionnary.
so instead of using a list, i made a dictionary. finally figured it out.
Finalized the analysis of numerical data.